Introduce WithSuspense component
