+++ site/frontend/src/pages/graphs.ts
@@ -1,5 +1,8 @@
 import Graphs from "./graphs/page.vue";
 import {createApp} from "vue";
+import WithSuspense from "../components/with-suspense.vue";
 
-const app = createApp(Graphs);
+const app = createApp(WithSuspense, {
+  component: Graphs
+});
 app.mount("#app");
